Full Emitters
Emitters are more or less done (I could add a function that makes the emitter centerOffset orbit and probably will eventually)
Angels now utilize the spinner and fire 4 projectiles in a spinning fashion.
Only basic projectiles work for the Emitter, but given that I'm about to redo the entire projectile creation system and likely the rest of the projectiles, I see no need to make them work until all of that happens.